logo, crl, data
Updated logo that was already there to have copyright with link to Co. site next to it.
Data was already in the feather file as discovered by JD.
Updated sliders in UI.R lines 231 and 307 to end in 2017 rather than 2016 (tried to set this to grab max, had trouble with factor.  Found ways to grab 2017 but got errors indicating the setting of factor was needed for other parts of the app).  Tested app briefly to make sure parts I touched still worked.  Did not deploy; will deploy after JH checks these in.